refactor: extract shared root command builder into internal/rootcmd
Both cmd/gcx/root and gcxlib/internal/embed duplicated the entire root
cobra command construction (PersistentPreRun, flags, subcommand
registration, help setup, agent annotations). Extract the shared logic
into internal/rootcmd.New() with an Options struct that accepts extra
commands and an optional provider list.

cmd/gcx/root passes dev.Command() via ExtraCommands while
gcxlib/internal/embed calls New() with nil options, preserving the
critical constraint that the embed path never imports cmd/gcx/dev.
